Changed dataTypes of APIs from UINT to int
Conflicts:
	MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/STM32F10x/DeviceCode/drivers/Samraksh_SPOT/Samraksh_SPOT.featureproj
	MicroFrameworkPK_v4_3/Samraksh/Interops/Samraksh_SPOT/Samraksh_SPOT/NativeStubs/Samraksh_SPOT.featureproj
	MicroFrameworkPK_v4_3/Samraksh/Interops/Samraksh_SPOT/Samraksh_SPOT/obj/Debug/Samraksh_SPOT.dll
	MicroFrameworkPK_v4_3/Samraksh/Interops/Samraksh_SPOT/Samraksh_SPOT/obj/Debug/Samraksh_SPOT.pdb

+++ MicroFrameworkPK_v4_3/DeviceCode/Include/Samraksh/DataStore/Types.h
@@ -7,7 +7,9 @@ typedef unsigned short  ushort;
 typedef unsigned char   uchar;
 
 
-typedef short RECORD_ID;
+//AnanthAtSamraksh - modified on 3/7/2014 - to store more recordIDs than is possible with short
+//typedef short RECORD_ID;
+typedef unsigned int RECORD_ID;
 
 #endif
